Catching exception when Authorization header is empty
Previously, if browser would send a:
  Authorized: Token
(where Token can be another auth token) or simply:
  Authorized: <blank>
http header, the split() method raises a ValueError resulting in a 500 error.
This errenous header should (probably) be handled as if it was not there or
otherwise invalid, catching the error and doing nothing does just that.